Updated test cases to have proper file lay out
